+++ datagrowth/configuration/__init__.py
@@ -1,7 +1,12 @@
 from datagrowth.configuration.types import (ConfigurationType, ConfigurationProperty, ConfigurationNotFoundError,
                                             create_config, register_defaults)
-from datagrowth.configuration.fields import ConfigurationField, ConfigurationFormField
 from datagrowth.configuration.serializers import load_config, DecodeConfigAction
+
+# Legacy imports
+try:
+    from datagrowth.django.fields import ConfigurationField
+except ImportError:
+    pass
 
 
 DATAGROWTH_CONFIGURATION = ConfigurationType()