+++ datagrowth/django/fields.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from datagrowth.django.configuration.field import ConfigurationField