+++ CHANGELOG.md
@@ -14,6 +14,7 @@ v0.22
 This update introduces LLM tools to the Datagrowth package and separates the package from its Django dependency. Being able to use the package outside of Django helps re-use in Jupyter for instance. It also makes it easier to implement snapshot testing, where snapshots no longer have to go through the Django fixture mechanisms.
 
 * To install Datagrowth you need to add to ``datagrowth.django.apps.DatagrowthConfig`` to ``INSTALLED_APPS``.
+* ``ConfigurationFormField`` is no longer exposed as a form to be used outside of the admin for ``ConfigurationField``.
 
 
 v0.21